requirements.txt
asttokens==2.0.5
attrs==21.2.0
bitstring==3.1.9
black==21.9b0
click==8.0.1
colorama==0.4.4
executing==0.8.1
flake8==3.9.2
icecream==2.1.1
iniconfig==1.1.1
mccabe==0.6.1
mypy-extensions==0.4.3
numpy==1.21.2
packaging==21.0
pathspec==0.9.0
Pillow==8.3.2
platformdirs==2.3.0
pluggy==1.0.0
py==1.10.0
pycodestyle==2.7.0
pyflakes==2.3.1
Pygments==2.10.0
pyparsing==2.4.7
pytest==6.2.5
regex==2021.8.28
scipy==1.7.1
six==1.16.0
toml==0.10.2
tomli==1.2.1
typing-extensions==3.10.0.2
